chore: moved hooks that needs use-mutation to use mutation
