--- src/api/authorize/use-get-settings.tsx
@@ -1,11 +1,0 @@
-import { useAuthorizeQuery } from '../../base';
-import { TSocketQueryOptions } from '../../base/use-query';
-
-export const useGetSettings = ({ ...props }: Omit<TSocketQueryOptions<'get_settings'>, 'name'> = {}) => {
-    const { data, ...rest } = useAuthorizeQuery({ name: 'get_settings', ...props });
-
-    return {
-        data: data?.get_settings,
-        ...rest,
-    };
-};
